service/requirements.txt
# Core Dependencies
alpaca-py==0.23.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
websockets==12.0

# Data & Storage
cryptography==42.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.0.3  # Required for SQLAlchemy async
motor==3.3.2  # Async MongoDB driver (optional)

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Notifications
aiohttp[speedups]==3.9.1
slack-sdk==3.26.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
tenacity==8.2.3  # Retry library
aiodebug==2.3.0  # Async debugging
uvloop==0.19.0  # High-performance event loop

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
mypy==1.8.0
ruff==0.1.14
tabulate==0.9.0  # For CLI table formatting

